[FIR] hacky fix lazy resolve contract violation for deprecation of enum entries
Enum entries can be used as types in Kotlin (even it's a compilation error)
To use the classifier as a type, we need to find if it's hidden or not.
The deprecations for classes are calculated on COMPILER_REQUIRED_ANNOTATION phase, and that's okay as it goes before the TYPES phase.
For enum entries, the deprecations are calculated on STATUS phase which goes after TYPES phase.
This is incorrect as we cannot jump from lower phase to upper phase (see ^KT-5489)

The hack ignores such deprecation search for enum entires.

hack for ^KT-55723