Reflection: create synthetic classes instead of throwing UOE
Another option would be to throw exception at the point where the
`::class` expression is used, if the expression's type on the left-hand
side is a synthetic class. But we can't really do that since it'll
affect performance of most `<expression>::class` expressions.

So, construct a fake synthetic class instead, without any members except
equals/hashCode/toString, and without any non-trivial modifiers. It kind
of contradicts the general idea that kotlin-reflect presents anything
exactly the same as the compiler sees it, but arguably it's worth it to
avoid unexpected exceptions like in KT-41373.

 #KT-41373 Fixed