[FIR] Fix K2 behavior according to RULES1
The compiler should only report diagnostics for
comparisons over builtins and identity-less types,
other incompatibilities should be reported
via inspections.

`SENSELESS_COMPARISON` is not reported in
`nullabilitySmartcastWhenNullability`,
because the non-smartcast types `Int? <:> Nothing?` should not lead(*)
to reporting a diagnostic, and the types after
the smartcast are `Nothing? <:> Nothing?` which also should not lead(*)
to a diagnostic.

(*) - according to RULES1 for primitives

It's ok that in `equalityChecksOnIntegerTypes`
instead of `EQUALITY_NOT_APPLICABLE_WARNING` we get
`EQUALITY_NOT_APPLICABLE`, because
`ProperEqualityChecksInBuilderInferenceCalls`
is already active by default.